refactor: modularize SyncHub screen into separate components
This commit refactors the `SyncHubScreen` by decomposing its monolithic structure into smaller, reusable UI components and moving them into a dedicated `ui.synchub` subpackage. This reorganization improves maintainability and follows better architectural patterns for the starter pack feature.

- **Component Extraction**:
    - Extracted `CatalogPackageCard` into its own file to handle standard package catalog items with status indicators and action buttons.
    - Extracted `HeroPackageCard` to manage the featured collection display, including its specific styling and security badges.
    - Extracted `GlowSyncTopAppBar` to encapsulate the search functionality and navigation header.
- **Package Reorganization**:
    - Relocated the primary `SyncHubScreen` to the `com.zoewave.probase.kocolor.features.starterpack.ui.synchub` package.
    - Updated `KoColorNavEntryProvider` imports to reflect the new package hierarchy.
- **Development Improvements**:
    - Added `@Preview` functions to all newly created component files to facilitate isolated UI testing and design iteration.
